requirements.txt
pymongo
python-dotenv
django-cors-headers
django-extensions
pandas
matplotlib
scikit-learn
openpyxl

